requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
billiard==4.2.1
bitarray==3.4.0
celery==5.5.2
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
chardet==5.2.0
charset-normalizer==3.4.2
ckzg==2.1.1
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
coverage==7.8.0
cryptography==44.0.3
cytoolz==1.0.1
daphne==4.2.0
dj-database-url==2.3.0
Django==5.2.1
django-cors-headers==4.7.0
django-filter==25.1
django-redis==5.4.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
Faker==37.3.0
frozenlist==1.6.0
gunicorn==23.0.0
hexbytes==1.3.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
iniconfig==2.1.0
kombu==5.5.3
msgpack==1.1.0
multidict==6.4.3
numpy==2.2.5
packaging==25.0
pandas==2.2.3
parsimonious==0.10.0
Pillow==9.5.0
pluggy==1.5.0
prompt_toolkit==3.0.51
propcache==0.3.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.9.0
pyOpenSSL==25.1.0
pytest==8.3.5
pytest-cov==6.1.1
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pyunormalize==16.0.0
redis==6.1.0
regex==2024.11.6
reportlab==4.4.1
requests==2.32.3
rlp==4.1.0
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.3
toolz==1.0.0
Twisted==24.11.0
txaio==23.1.1
types-requests==2.32.0.20250328
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
web3==7.11.1
websockets==15.0.1
whitenoise==6.9.0
yarl==1.20.0
zope.interface==7.2
